requirements.txt
pandas==2.2.3
openpyxl==3.1.5
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Jinja2==3.1.4
reportlab==4.2.5